Rename base_ai to base
