+++ providers/common/ai/docs/changelog.rst
@@ -58,7 +58,7 @@ Features
 
 .. For the next release: direct callers of :meth:`~airflow.providers.common.ai.hooks.pydantic_ai.PydanticAIHook.create_agent`
    and :meth:`~airflow.providers.common.ai.hooks.pydantic_ai.PydanticAIHook.run_agent` must use
-   :class:`~airflow.providers.common.ai.hooks.base_ai.AgentRunRequest` instead of keyword arguments.
+   :class:`~airflow.providers.common.ai.hooks.base.AgentRunRequest` instead of keyword arguments.
    DAG authors using :class:`~airflow.providers.common.ai.operators.agent.AgentOperator`,
    ``@task.agent``, and the other LLM operators are unaffected.
