+++ providers/common/ai/provider.yaml
@@ -64,7 +64,7 @@ integrations:
 hooks:
   - integration-name: Common AI
     python-modules:
-      - airflow.providers.common.ai.hooks.base_ai
+      - airflow.providers.common.ai.hooks.base
   - integration-name: Pydantic AI
     python-modules:
       - airflow.providers.common.ai.hooks.pydantic_ai
