+++ providers/common/ai/src/airflow/providers/common/ai/hooks/base.py
